requirements.txt
anyio==4.10.0
certifi==2025.8.3
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
peewee==3.18.2
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.1
python-telegram-bot==22.3
rsa==4.9.1
sniffio==1.3.1
Telethon==1.40.0
typing_extensions==4.14.1